src/requirements.txt
asgiref==3.8.1
attrs==24.2.0
blinker==1.8.2
boto3==1.35.54
botocore==1.35.54
click==8.1.7
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
Flask==3.0.3
greenlet==3.1.1
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
packaging==24.1
pika==1.3.2
pillow==11.0.0
pluggy==1.5.0
postgres==4.0
psycopg2-binary==2.9.10
psycopg2-pool==1.2
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.1
s3transfer==0.10.3
six==1.16.0
SQLAlchemy==2.0.36
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.1.2
